requirements.txt
# VibeSafe — Python 의존성
# 설치: pip install -r requirements.txt

# ─── 핵심 스캐닝 ─────────────────────────────────────────
semgrep>=1.60.0              # SAST 엔진
pip-audit>=2.7.0             # Python SCA

# ─── AI 수정 제안 ─────────────────────────────────────────
anthropic>=0.25.0            # Claude API (auto_fix_generator)

# ─── 데이터베이스 ─────────────────────────────────────────
psycopg2-binary>=2.9.9       # PostgreSQL
redis>=5.0.0                 # Redis 캐시

# ─── Celery 워커 ──────────────────────────────────────────
celery>=5.3.0                # 비동기 태스크 큐
boto3>=1.34.0                # AWS S3 (MinIO 포함)

# ─── 유틸리티 ─────────────────────────────────────────────
python-dotenv>=1.0.0         # .env 파일 로드
